+++ iOS/Projects/Features/Login/Sources/LoginScene/LoginViewController.swift
@@ -1,0 +1,30 @@
+//
+//  LoginViewController.swift
+//  LoginFeature
+//
+//  Created by 안종표 on 11/28/23.
+//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
+//
+
+import UIKit
+
+public final class LoginViewController: UIViewController {
+  
+  private let viewModel: LoginViewModelRepresentable
+
+  public init(viewModel: LoginViewModelRepresentable) {
+    self.viewModel = viewModel
+    super.init(nibName: nil, bundle: nil)
+  }
+
+  @available(*, unavailable)
+  required init?(coder _: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+
+  override public func viewDidLoad() {
+    super.viewDidLoad()
+  }
+  
+  
+}
