+++ iOS/Projects/Features/Login/Sources/LoginScene/ViewModel/Protocol/LoginViewModelRepresentable.swift
@@ -1,0 +1,13 @@
+//
+//  LoginViewModelRepresentable.swift
+//  LoginFeature
+//
+//  Created by 안종표 on 11/28/23.
+//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
+//
+
+import Foundation
+
+public protocol LoginViewModelRepresentable {
+  
+}
