--- iOS/Projects/Features/Login/Sources/TempScene/TempViewController.swift
@@ -1,31 +1,0 @@
-import UIKit
-
-
-public final class TempViewController: UIViewController {
-
-  // MARK: Properties
-  // 여기에 UI를 제외한 나머지 프로퍼티를 설정하세요.
-
-  private let viewModel: TempViewModelRepresentable
-
-  // MARK: UI Components
-  // 여기에 View 프로퍼티를 설정하세요.
-
-  // MARK: Initializations
-
-  public init(viewModel: TempViewModelRepresentable) {
-    self.viewModel = viewModel
-    super.init(nibName: nil, bundle: nil)
-  }
-
-  @available(*, unavailable)
-  required init?(coder _: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  // MARK: Life Cycles
-
-  override public func viewDidLoad() {
-    super.viewDidLoad()
-  }
-}
