--- iOS/Projects/Features/Login/Sources/TempScene/TempViewViewModel.swift
@@ -1,36 +1,0 @@
-import Combine
-
-public struct TempViewModelInput {}
-
-public typealias TempViewModelOutput = AnyPublisher<TempState, Never>
-
-public enum TempState {
-  case idle
-}
-
-public protocol TempViewModelRepresentable {
-  func transform(input: TempViewModelInput) -> TempViewModelOutput
-}
-
-public final class TempViewModel {
-  // MARK: Properties
-
-  private var subscriptions: Set<AnyCancellable> = []
-
-  // MARK: Initializations
-
-  public init() {}
-}
-
-extension WorkoutSummaryViewModel: TempViewModelRepresentable {
-  public func transform(input: TempViewModelInput) -> TempViewModelOutput {
-    for subscription in subscriptions {
-      subscription.cancel()
-    }
-    subscriptions.removeAll()
-
-    let initialState: WorkoutSummaryViewModelOutput = Just(.idle).eraseToAnyPublisher()
-
-    return initialState
-  }
-}
